[ats_coverage] Refactor based on SOLID principles, updated unit tests, docs, added support for content generators
